Changed navigation
Removed book path and used dashboard, updated links in nav bar
