--- pages/books/index.js
@@ -1,9 +1,0 @@
-import BookDashboard from "@/components/book/book-dashboard";
-
-export default function Books() {
-	return (
-		<>
-			<BookDashboard />
-		</>
-	);
-}
