+++ pages/dashboard/index.js
@@ -1,0 +1,9 @@
+import BookDashboard from "@/components/book/book-dashboard";
+
+export default function Books() {
+	return (
+		<>
+			<BookDashboard />
+		</>
+	);
+}
